requirements.txt
pygame-ce
pygame-menu-ce
PyGLM
moderngl
numpy
numba
opensimplex
coloredlogs
